+++ feature/daily/data/src/main/java/com/example/betteryou/feature/daily/data/remote/model/product/ProductDto.kt
@@ -7,7 +7,7 @@ data class ProductDto(
     val id: Int,
     val name: String,
     val photo: String,
-    val calories: Int,
+    val calories: Double,
     val protein: Double,
     val carbs: Double,
     val fat: Double,
