+++ feature/daily/domain/src/main/java/com/example/betteryou/feature/daily/domain/model/Intake.kt
@@ -1,9 +1,9 @@
 package com.example.betteryou.feature.daily.domain.model
 
 data class Intake(
-    val dailyCalories:Int,
-    val protein: Int,
-    val fats: Int,
-    val carbs: Int,
+    val dailyCalories:Double,
+    val protein: Double,
+    val fats: Double,
+    val carbs: Double,
     val water: Double
 )
