+++ feature/daily/domain/src/main/java/com/example/betteryou/feature/daily/domain/model/Nutrient.kt
@@ -1,9 +1,9 @@
 package com.example.betteryou.feature.daily.domain.model
 
 data class Nutrient(
-    val protein: Int,
-    val fats: Int,
-    val carbs: Int,
+    val protein: Double,
+    val fats: Double,
+    val carbs: Double,
     val water:Double,
-    val dailyCalories:Int
+    val dailyCalories:Double
 )
