+++ feature/daily/domain/src/main/java/com/example/betteryou/feature/daily/domain/repository/user_daily_product/UserDailyProductRepository.kt
@@ -12,5 +12,5 @@ interface UserDailyProductRepository {
 
     suspend fun clearOldProducts(userId: String)
 
-    suspend fun deleteProductById(userId: String,productId: Int)
+    suspend fun deleteProductById(userId: String, productId: Long)
 }
