+++ feature/daily/domain/src/main/java/com/example/betteryou/feature/daily/domain/usecase/user_daily_product/DeleteProductByIdUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class DeleteProductByIdUseCase @Inject constructor(
     private val productRepository: UserDailyProductRepository,
 ) {
-    suspend operator fun invoke(productId: Int,userId:String) {
+    suspend operator fun invoke(productId: Long, userId: String) {
         productRepository.deleteProductById(userId,productId)
     }
 }
